Completed server menu
Internet connection work properly
Need more testing

+++ src/game/connectMenu/targetConnect.cpp
@@ -136,7 +136,7 @@ void TargetConnect::tryConnect() {
     memcpy(basePort, portTextCorrected, sizeof(basePort));
     // Trying connect at specified address
     Destination dest{IPField.getString(), (Uint16)SDL_atoi(portTextCorrected)};
-    internet.sendFirst(dest, {ConnectionCode::Init, 1});
+    internet.sendFirst(dest, {ConnectionCode::Init, Uint8(1)});
 }
 
 void TargetConnect::blit() const {
